move FirecloudGroupModel out of openapi spec
