+++ src/main/java/bio/terra/service/duos/DuosService.java
@@ -5,8 +5,8 @@ import bio.terra.common.FutureUtils;
 import bio.terra.model.DuosFirecloudGroupModel;
 import bio.terra.model.DuosFirecloudGroupsSyncResponse;
 import bio.terra.model.ErrorModel;
-import bio.terra.model.FirecloudGroupModel;
 import bio.terra.model.RepositoryStatusModelSystems;
+import bio.terra.service.auth.iam.FirecloudGroupModel;
 import bio.terra.service.auth.iam.IamRole;
 import bio.terra.service.auth.iam.IamService;
 import bio.terra.service.duos.exception.DuosDatasetBadRequestException;
