+++ src/test/java/bio/terra/service/duos/DuosServiceTest.java
@@ -25,8 +25,8 @@ import bio.terra.common.category.Unit;
 import bio.terra.common.fixtures.DuosFixtures;
 import bio.terra.model.DuosFirecloudGroupModel;
 import bio.terra.model.ErrorModel;
-import bio.terra.model.FirecloudGroupModel;
 import bio.terra.model.RepositoryStatusModelSystems;
+import bio.terra.service.auth.iam.FirecloudGroupModel;
 import bio.terra.service.auth.iam.IamRole;
 import bio.terra.service.auth.iam.IamService;
 import bio.terra.service.auth.iam.exception.IamForbiddenException;
